fix(cutover): 收口六路由错误合同
- 严格解码列表查询参数与锁定枚举
- 分离任务和评估版本冲突及当前版本
- 清除活动任务冲突的错误Owner归因

+++ pms-module-cutover/src/main/java/cn/iocoder/yudao/module/pms/cutover/controller/admin/taskv2/CutoverTaskErrorData.java
@@ -2,6 +2,6 @@ package cn.iocoder.yudao.module.pms.cutover.controller.admin.taskv2;
 
 /** F-CUT-002局部错误合同。不可得的当前版本保持null。 */
 public record CutoverTaskErrorData(String category, String reasonCode, String recoveryAction,
-                                   String ownerContext, Long currentTaskVersion,
-                                   Long currentAssessmentVersion, Long currentOwnerVersion) {
+                                   String ownerContext, Integer currentTaskVersion,
+                                   Integer currentAssessmentVersion, Long currentOwnerVersion) {
 }
